+++ frontend/src/app/app-routing.module.ts
@@ -9,6 +9,7 @@ import { SignupComponent } from './auth/signup/signup.component';
 import { ProductosComponent } from './pages/productos/productos.component';
 import { PagoComponent } from './pages/pago/pago.component';
 import { PedidosComponent } from './pages/pedidos/pedidos.component';
+import { authGuard } from './guards/auth.guard';
 
 const routes: Routes = [
   {path: '', redirectTo:'/home', pathMatch:'full'},
@@ -18,9 +19,9 @@ const routes: Routes = [
   {path: 'carta', component:CartaComponent},
   {path: 'login', component:LoginComponent},
   {path: 'signup', component:SignupComponent},
-  {path: 'productos', component:ProductosComponent},
+  {path: 'productos', component:ProductosComponent, canActivate:[authGuard]},
   {path: 'pago', component:PagoComponent},
-  {path: 'pedidos', component:PedidosComponent}
+  {path: 'pedidos', component:PedidosComponent, canActivate:[authGuard]}
 ];
 
 @NgModule({
