--- frontend/src/app/auth/signup/signup.component.spec.ts
@@ -1,21 +1,0 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-import { SignupComponent } from './signup.component'
-
-
-describe('SignupComponent', () => {
-  let component: SignupComponent;
-  let fixture: ComponentFixture<SignupComponent>;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({
-      declarations: [SignupComponent]
-    });
-    fixture = TestBed.createComponent(SignupComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
