+++ frontend/src/app/guards/auth.guard.ts
@@ -1,0 +1,14 @@
+import { inject } from '@angular/core';
+import { Router, CanActivateFn } from '@angular/router';
+
+export const authGuard: CanActivateFn = (route, state) => {
+  const router = inject(Router);
+  const isAdmin = localStorage.getItem('is_staff');
+
+  if (!isAdmin) {
+    router.navigate(['/login']);
+    return false;
+  }
+
+  return true;
+};
