+++ frontend/src/app/services/auth/authRequest.ts
@@ -1,3 +1,8 @@
+export interface LoginRequest {
+    email:string,
+    password:string
+}
+
 export interface SignupRequest {
     email:string,
     username:string,
