--- frontend/src/app/services/auth/login.service.ts
@@ -1,16 +1,0 @@
-import { Injectable } from '@angular/core';
-import { LoginRequest } from './loginRequest';
-import { HttpClient } from '@angular/common/http';
-import { Observable } from 'rxjs';
-
-@Injectable({
-  providedIn: 'root',
-})
-export class LoginService {
-  private url: string = 'http://127.0.0.1:8000/api/auth/login/';
-  constructor(private http: HttpClient) {}
-
-  login(credentials: LoginRequest): Observable<any> {
-    return this.http.post(this.url, credentials);
-  }
-}
