--- frontend/src/app/services/auth/loginRequest.ts
@@ -1,4 +1,0 @@
-export interface LoginRequest {
-    email:string,
-    password:string
-}
