--- frontend/src/app/services/auth/signup.service.spec.ts
@@ -1,16 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { SignupService } from './signup.service';
-
-describe('SignupService', () => {
-  let service: SignupService;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({});
-    service = TestBed.inject(SignupService);
-  });
-
-  it('should be created', () => {
-    expect(service).toBeTruthy();
-  });
-});
